splitting up what has become an enormous file
